requirements.txt
streamlit
pandas
python-dotenv
Pillow